fixed encode and findInDictionary functions

+++ compression/compression/lz.h
@@ -16,4 +16,4 @@ typedef struct dictionaryElem
 
 returnCode lz78_encode(FILE* in, FILE* out);
 returnCode lz78_decode(FILE* in, FILE* out);
-unsigned short findInDictionary(dictionaryElem* dictionary,unsigned short length,unsigned short startIndex, char elem);
+unsigned short findInDictionary(dictionaryElem* dictionary,int length,unsigned short startIndex, char elem);
